Swapping out boost asio for gRPC. This isn't done yet
