+++ src/protos/ristretto.proto
@@ -1,0 +1,15 @@
+syntax = "proto3"
+
+package ristretto;
+
+service Greeter {
+    rpc sayHello (HelloRequest) return (HelloReply) {}
+}
+
+message HelloRequest {
+    string name = 1
+}
+
+message HelloReply {
+    string message = 1
+}
